create bombs module
