+++ vue/src/store/field/field.js
@@ -1,5 +1,6 @@
-import { ALL_ORIENTATIONS } from './pieces'
-import { DECREASE_TIMER_VALUE_DEFAULT, INCREASE_TIMER_VALUE_DEFAULT } from './game'
+import { ALL_ORIENTATIONS } from '../pieces'
+import { DECREASE_TIMER_VALUE_DEFAULT, INCREASE_TIMER_VALUE_DEFAULT } from '../game'
+import { OBJECTS } from './objects'
 
 const MUTATIONS = {
     CHANGE_FIELD_SIZE: 'CHANGE_FIELD_SIZE',
@@ -10,20 +11,6 @@ const MUTATIONS = {
     SET_FIELD: 'SET_FIELD',
     CLEAR_FIELD: 'CLEAR_FIELD',
     SET_CURRENT_FIGURE_COORDS: 'SET_CURRENT_FIGURE_COORDS',
-    ADD_BOMB: 'ADD_BOMB',
-    REMOVE_BOMB: 'REMOVE_BOMB',
-    CLEAR_BOMBS: 'CLEAR_BOMBS',
-    UPDATE_BOMB_POSITION: 'UPDATE_BOMB_POSITION',
-}
-
-export const OBJECTS = {
-  NONE: 0,
-  CENTRAL_CUBE: 1,
-  EXTERNAL_FIGURE: 2,
-  ATTACHED_CUBE: 3,
-  BLACK_BOMB: 11,
-  RED_BOMB: 12,
-  GREEN_BOMB: 13
 }
 
 export default {
@@ -36,7 +23,6 @@ export default {
         currentPiece: null,
         currentFigureCoords: null,
         lastSide: null,
-        bombs: []
     }
   },
   getters: {
@@ -45,7 +31,6 @@ export default {
     isGameActive: (state) => state.gameActive,
     getCurrentPiece: (state) => state.currentPiece,
     getCurrentFigureCoords: state => state.currentFigureCoords,
-    getBombs: (state) => state.bombs
   },
   mutations: {
     [MUTATIONS.CHANGE_FIELD_SIZE]: (state, newSize) => {
@@ -94,21 +79,6 @@ export default {
     },
     [MUTATIONS.SET_CURRENT_FIGURE_COORDS]: (state, value) => {
         state.currentFigureCoords = value
-    },
-    [MUTATIONS.ADD_BOMB]: (state, bomb) => {
-        state.bombs.push(bomb)
-    },
-    [MUTATIONS.REMOVE_BOMB]: (state, index) => {
-        state.bombs.splice(index, 1)
-    },
-    [MUTATIONS.CLEAR_BOMBS]: (state) => {
-        state.bombs = []
-    },
-    [MUTATIONS.UPDATE_BOMB_POSITION]: (state, { index, x, y }) => {
-        if (state.bombs[index]) {
-            state.bombs[index].x = x
-            state.bombs[index].y = y
-        }
     },
   },
   actions: {
@@ -298,15 +268,18 @@ export default {
         } else {
             const updatedPiece = { ...piece, x, y }
             store.commit(MUTATIONS.SET_CURRENT_PIECE, updatedPiece)
-            const bombs = store.state.bombs
+            const bombs = store.rootGetters['bombs/getBombs']
             store.dispatch('drawPieceOnField').then(
                 () => store.dispatch('checkFigureAttachment')
             ).then(
                 () => {
-                    for (let i = 0; i < bombs.length; i++) {
-                        const bomb = bombs[i]
+                    bombs.forEach((bomb, i) => {
                         store.dispatch('checkCrash', {index: i, x: bomb.x, y: bomb.y})
-                    }
+                    })
+                    // for (let i = 0; i < bombs.length; i++) {
+                    //     const bomb = bombs[i]
+                    //     store.dispatch('checkCrash', {index: i, x: bomb.x, y: bomb.y})
+                    // }
                 }
             )
         }
@@ -504,136 +477,5 @@ export default {
           queue.push({x: x, y: y})
           store.dispatch('cube/addPiece', {x: x - centralCubePosition.x, y: y - centralCubePosition.y}, { root: true })
     },
-    spawnBomb: (store) => {
-        if (!store.state.gameActive) return
-        if (store.state.bombs.length >= 3) return
-
-        const rand = Math.random()
-        let color
-        if (rand <= 0.1) {
-            color = OBJECTS.GREEN_BOMB
-        } else {
-            color = rand < 0.55 ? OBJECTS.BLACK_BOMB : OBJECTS.RED_BOMB
-        }
-
-        const side = Math.floor(Math.random() * 4)
-        const fieldSize = store.state.size
-        let x, y, direction
-
-        const getPosition = (minPossible, maxPossible) => {
-            if (maxPossible - minPossible >= 2) {
-                const newMin = minPossible + 1
-                const newMax = maxPossible - 1
-                return Math.floor(Math.random() * (newMax - newMin + 1)) + newMin
-            }
-            return Math.floor(Math.random() * (maxPossible - minPossible + 1)) + minPossible
-        }
-
-        switch (side) {
-            case 0:
-                x = getPosition(0, fieldSize - 1)
-                y = -1
-                direction = 0
-                break
-            case 1:
-                x = fieldSize
-                y = getPosition(0, fieldSize - 1)
-                direction = 1
-                break
-            case 2:
-                x = getPosition(0, fieldSize - 1)
-                y = fieldSize
-                direction = 2
-                break
-            case 3:
-                x = -1
-                y = getPosition(0, fieldSize - 1)
-                direction = 3
-                break
-        }
-        const bomb = { x, y, direction, color }
-        store.commit(MUTATIONS.ADD_BOMB, bomb)
-    },
-
-    moveBombs: (store) => {
-        if (!store.state.gameActive) return
-        const fieldSize = store.state.size
-        const bombs = store.state.bombs
-        for (let i = 0; i < bombs.length; i++) {
-            const bomb = bombs[i]
-            let { x, y, direction, color } = bomb
-            const oldX = x, oldY = y
-
-            switch (direction) {
-                case 0: y++; break
-                case 1: x--; break
-                case 2: y--; break
-                case 3: x++; break
-            }
-
-            if (x < 0 || x >= fieldSize || y < 0 || y >= fieldSize) {
-                if (oldX >= 0 && oldX < fieldSize && oldY >= 0 && oldY < fieldSize) {
-                    store.commit(MUTATIONS.SET_NUMBER, { position: {x: oldX, y: oldY}, number: OBJECTS.NONE })
-                }
-                store.commit(MUTATIONS.REMOVE_BOMB, i)
-                continue
-            }
-
-            const cell = store.state.field[y][x]
-            if (cell === OBJECTS.CENTRAL_CUBE || cell === OBJECTS.ATTACHED_CUBE) {
-                store.dispatch('handleBombCollision', { bomb: { ...bomb, x: oldX, y: oldY }, index: i })
-                continue
-            }
-
-            if (oldX >= 0 && oldX < fieldSize && oldY >= 0 && oldY < fieldSize) {
-                store.commit(MUTATIONS.SET_NUMBER, { position: {x: oldX, y: oldY}, number: OBJECTS.NONE })
-            }
-            store.dispatch('checkCrash', {index: i, x, y}).then(
-                isCrash => {
-                    if (!isCrash) {
-                        store.commit(MUTATIONS.SET_NUMBER, { position: {x, y}, number: color })
-                        store.commit(MUTATIONS.UPDATE_BOMB_POSITION, { index: i, x, y })
-                    }
-                }
-            )
-        }
-    },
-    checkCrash: (store, {index, x, y}) => {
-        const piece = store.state.currentPiece
-
-        let isCrash = false
-
-        for (let r = 0; r < piece.shape.length && !isCrash; r++) {
-            for (let c = 0; c < piece.shape[0].length && !isCrash; c++) {
-                if (piece.shape[r][c] === 1) {
-                    const nx = piece.x + c
-                    const ny = piece.y + r
-                    isCrash ||= x == nx && y == ny
-                }
-            }
-        }
-        if (isCrash) {
-            store.dispatch('spawnPiece')
-            store.commit(MUTATIONS.REMOVE_BOMB, index)
-        }
-        return isCrash
-    },
-    handleBombCollision: (store, { bomb, index }) => {
-        const color = bomb.color
-        store.commit(MUTATIONS.REMOVE_BOMB, index)
-        if (bomb.x >= 0 && bomb.x < store.state.size && bomb.y >= 0 && bomb.y < store.state.size) {
-            store.commit(MUTATIONS.SET_NUMBER, { position: {x: bomb.x, y: bomb.y}, number: OBJECTS.NONE })
-        }
-
-        if (color === OBJECTS.BLACK_BOMB) {
-            store.dispatch('cube/bombClearAllAttachedPieces', null, { root: true }).then(count => {
-            store.dispatch('game/addScore', -count, { root: true })
-            })
-        } else if (color === OBJECTS.RED_BOMB) {
-            store.dispatch('game/updateTimer', { isIncrease: false, decreaseValue: 30, increaseValue: 0 }, { root: true })
-        } else if (color === OBJECTS.GREEN_BOMB) {
-            store.dispatch('game/updateTimer', { isIncrease: true, decreaseValue: 0, increaseValue: 10 }, { root: true })
-        }
-    }
   }
 }
