+++ vue/src/store/index.js
@@ -1,8 +1,9 @@
 import { createStore } from 'vuex'
 import list from './list'
-import field from './field'
+import field from './field/field'
 import game from './game'
 import cube from './cube'
+import bombs from './field/bombs'
 
 const MUTATIONS = {
   INCREMENT: 'INCREMENT',
@@ -43,6 +44,7 @@ export default createStore({
     list,
     field,
     game,
-    cube
+    cube,
+    bombs
   }
 })
